+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Expect active development and potentially significant breaking changes in the `0.x` track. We'll try to be diligent about releasing a `1.0` version in a timely fashion (ideally within 3 to 6 months), to signal the start of a more stable API.
 
 ### vNEXT
-- ...
+- Add `QueryManager` as an index export [PR #1269](https://github.com/apollographql/apollo-client/pull/1269)
 
 
 ### 0.8.2